requirements.txt
requests
beautifulsoup4
spotipy
tqdm
python-dotenv